% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{job_force_stop}
\alias{job_force_stop}
\alias{job_force_start}
\alias{flows_get}
\title{job_force_stop}
\usage{
job_force_stop(flow_name, job_name)

job_force_start(flow_name, job_name)

flows_get()
}
\arguments{
\item{flow_name}{}

\item{job_name}{}
}
\description{
job_force_stop
}
\section{Functions}{
\itemize{
\item \code{job_force_start()}: job_force_start

\item \code{flows_get()}: flows_get

}}
