% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_server}
\alias{api_server}
\alias{api_job_start}
\alias{api_job_stop}
\alias{api_flows_get}
\title{Tessiflow API server functions}
\usage{
api_job_start(flow_name, job_name)

api_job_stop(flow_name, job_name)

api_flows_get()
}
\arguments{
\item{flow_name}{character workflow name}

\item{job_name}{character job name}
}
\description{
Tessiflow API server functions
}
\section{Functions}{
\itemize{
\item \code{api_job_start()}: internal api function for forced starting a job

\item \code{api_job_stop()}: internal api function for forced stopping a job

\item \code{api_flows_get()}: internal api function for getting last run of jobs

}}
