% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_start}
\alias{api_start}
\title{api_start}
\usage{
api_start(working_dir = getwd(), docs = FALSE, debug = FALSE)
}
\arguments{
\item{working_dir}{character, working directory to start the API server in}

\item{docs}{boolean, passed on to \link[plumber:pr_run]{plumber::pr_run}}

\item{debug}{boolean, passed on to \code{plumber::pr_run}}
}
\description{
Start the plumber API for tessiflow and override the working directory and some \code{plumber} startup options
}
