% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_flow_job_name.R
\name{assert_flow_job_name}
\alias{assert_flow_job_name}
\title{assert_flow_job_name}
\usage{
assert_flow_job_name(flow_name, job_name)
}
\arguments{
\item{flow_name}{string flow name}

\item{job_name}{string job name}
}
\description{
assertions for flow and job name: both are characters of length 1 and both exist in flows
}
