% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_flow_job_name.R
\name{check_flow_job_name}
\alias{check_flow_job_name}
\alias{assert_flow_job_name}
\alias{test_flow_job_name}
\title{check_flow_job_name}
\usage{
check_flow_job_name(flow_name, job_name)

assert_flow_job_name(
  flow_name,
  job_name,
  .var.name = checkmate::vname(flow_name),
  add = NULL
)

test_flow_job_name(flow_name, job_name)
}
\arguments{
\item{flow_name}{\code{character} flow name}

\item{job_name}{\code{character} job name}

\item{.var.name}{\code{character} name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.}

\item{add}{\code{AssertCollection} collection to store assertion messages. See \code{AssertCollection.}}
}
\description{
check, assertion, and test for flow and job name: both are characters of length 1 and both exist in flows
}
\section{Functions}{
\itemize{
\item \code{assert_flow_job_name()}: errors if check is false

\item \code{test_flow_job_name()}: returns \code{FALSE} if check is false

}}
