% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handlers.R
\name{error_handler}
\alias{error_handler}
\alias{error_print}
\alias{error_email}
\alias{error_handler_factory}
\title{error_handler}
\usage{
error_handler(error)

error_print(error)

error_email(
  error,
  emails = config::get("tessiflow.email"),
  smtp = config::get("tessiflow.smtp")
)

error_handler_factory(fun)
}
\arguments{
\item{error}{error condition}

\item{emails}{list of email addresses, default is from \code{tessiflow.email} configuration variable; \code{from} address will be the first email}

\item{smtp}{list of smtp configuration data for \code{\link[mailR:send.mail]{mailR::send.mail}}}

\item{fun}{function to wrap}
}
\description{
functions for handling errors
}
\section{Functions}{
\itemize{
\item \code{error_handler}: selects error handler to use

\item \code{error_print}: prints an error message to the console

\item \code{error_email}: sends an error message to an email address

\item \code{error_handler_factory}: wrap a function in an error handler
}}

