% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_parse.R
\name{flow_to_data_table}
\alias{flow_to_data_table}
\title{flow_to_data_table}
\usage{
flow_to_data_table(flow)
}
\arguments{
\item{flow}{single workflow nested list generated by flows_parse}
}
\value{
flow as a data.table
}
\description{
parses a single flow from a yml file to a data.table
}
