% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_parse.R
\name{flows_auto_refresh}
\alias{flows_auto_refresh}
\title{flows_auto_refresh}
\usage{
flows_auto_refresh(flows_directory = config::get("tessiflow.d"))
}
\arguments{
\item{flows_directory}{directory containing the yml files to load}
}
\value{
updated data.table
}
\description{
Update flows data.table with new data from yml flows if they have changed since the last load
}
