% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_main.R
\name{flows_get_job}
\alias{flows_get_job}
\title{flows_get_job}
\usage{
flows_get_job(.flow_name, .job_name)
}
\arguments{
\item{.flow_name}{string workflow name}

\item{.job_name}{string job name}
}
\value{
list, one row of the flows table
}
\description{
Lookup the flow row in the local data.table and return it as a list
}
