% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_log.R
\name{flows_log_cleanup}
\alias{flows_log_cleanup}
\title{flows_log_cleanup}
\usage{
flows_log_cleanup()
}
\value{
invisibly
}
\description{
Cleans up jobs in the log that are marked as \code{Running} but no longer are.
}
