% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_log.R
\name{flows_log_get_create_time}
\alias{flows_log_get_create_time}
\title{flows_log_get_create_time}
\usage{
flows_log_get_create_time()
}
\value{
creation time of the log database
}
\description{
flows_log_get_create_time
}
