% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_log.R
\name{flows_log_get_last_run}
\alias{flows_log_get_last_run}
\title{flows_log_get_last_run}
\usage{
flows_log_get_last_run(flows, jobs)
}
\arguments{
\item{flows}{character vector of flows}

\item{jobs}{character vector of jobs}
}
\value{
matching rows from the log database for the last run of each flow, job pair
}
\description{
flows_log_get_last_run
}
