% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_log.R
\name{flows_log_open}
\alias{flows_log_open}
\alias{flows_log_close}
\alias{flows_log_create}
\title{flows_log_open}
\usage{
flows_log_open(flows_log_dir = config::get("tessiflow.log"))

flows_log_close()

flows_log_create()
}
\arguments{
\item{flows_log_dir}{directory where the SQLite database is stored}
}
\value{
invisible
}
\description{
flows_log_open
}
\section{Functions}{
\itemize{
\item \code{flows_log_close}: Tear down the SQL connection

\item \code{flows_log_create}: Create the SQLite log table
}}

