% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_log.R
\name{flows_log_upsert}
\alias{flows_log_upsert}
\title{flows_log_upsert}
\usage{
flows_log_upsert(table = "jobs", data)
}
\arguments{
\item{table}{string, right now only \code{jobs} is valid}

\item{data}{data.frame of data to update, rows will be matched on flow_name, job_name, and start_time}
}
\value{
invisibly
}
\description{
flows_log_upsert
}
