% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_log.R
\name{flows_log_write}
\alias{flows_log_write}
\title{flows_log_write}
\usage{
flows_log_write(table = "jobs", data)
}
\arguments{
\item{table}{string, right now only \code{jobs} is valid}

\item{data}{data.frame of data to write}
}
\value{
invisibly
}
\description{
flows_log_write
}
