% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_main.R
\name{flows_main_read_server}
\alias{flows_main_read_server}
\title{flows_main_read_server}
\usage{
flows_main_read_server(server)
}
\arguments{
\item{server}{Server object created by \code{base::socketServer}}
}
\value{
result of call
}
\description{
Reads from socket server and (safely) executes commands by checking the call name and converting arguments to strings.
}
