% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_parse.R
\name{flows_parse}
\alias{flows_parse}
\title{flows_parse}
\usage{
flows_parse(flows_directory = config::get("tessiflow.d"))
}
\arguments{
\item{flows_directory}{the directory containing the yml files to load}
}
\value{
named list of workflows as data.tables
}
\description{
Loads yml workflow files from \code{flows_directory} and returns a list of data.tables with the workflow data
}
