% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_parse.R
\name{flows_refresh}
\alias{flows_refresh}
\title{flows_refresh}
\usage{
flows_refresh(...)
}
\arguments{
\item{...}{extra arguments passed on to \code{flows_parse}}
}
\value{
updated data.table
}
\description{
Update flows data.table with new data from yml flows
}
