% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{flows_report_load}
\alias{flows_report_load}
\alias{flows_report_send}
\title{flows_report}
\usage{
flows_report_load()

flows_report_send()
}
\value{
nothing
}
\description{
flows_report
}
\section{Functions}{
\itemize{
\item \code{flows_report_load}: Pull process data from the SQLite database from the last 24hours.

\item \code{flows_report_send}: Send the flows report
}}

