% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_main.R
\name{flows_update_job}
\alias{flows_update_job}
\title{flows_update_job}
\usage{
flows_update_job(.flow_name, .job_name, data)
}
\arguments{
\item{.flow_name}{string workflow name}

\item{.job_name}{string job name}

\item{data}{list of data to update the flow with}
}
\value{
invisibly
}
\description{
flows_update_job
}
