% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_main.R
\name{flows_update_status}
\alias{flows_update_status}
\title{flows_update_status}
\usage{
flows_update_status(flows)
}
\arguments{
\item{flows}{data.table of flows}
}
\value{
data.table of flows with added columns:
\itemize{
\item prev_scheduled_run: POSIXct previous scheduled run time, generated by parse_cron from on.schedule
\item next_scheduled_run: POSIXct next scheduled run time, generated by parse_cron from on.schedule
\item last_start: POSIXct last start time, updated when task is run
\item last_end: POSIXct last end time, updated when task is completed or by flow_get_last_run
\item status: string (\code{Waiting},\code{Running},\code{Finished})
\item retval: integer return value, or NA if not yet \code{Finished}
}
}
\description{
flows_update_status
}
