% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_log.R
\name{job_log_write}
\alias{job_log_write}
\title{job_log_write}
\usage{
job_log_write(flow_name, job_name = NA, lines, console = FALSE)
}
\arguments{
\item{flow_name}{character workflow name}

\item{job_name}{character job name}

\item{lines}{character vector of lines to write to the file}

\item{console}{boolean whether to echo to the console, default is FALSE}
}
\description{
Simple logfile writer, with filename determined by the workflow. Timestamp and job name are prepended to each line before writing.
}
