% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_make_remote_expr}
\alias{job_make_remote_expr}
\title{job_make_remote_expr}
\usage{
job_make_remote_expr(
  env_vars = NULL,
  if_expr = NULL,
  run_expr = NULL,
  shell = "callr"
)
}
\arguments{
\item{env_vars}{named list of environment variables}

\item{if_expr}{deparsed if expression}

\item{run_expr}{deparsed run expression}

\item{shell}{string setting the shell, default is \code{callr}}
}
\value{
R expression
}
\description{
Creates an expression object that has local environment variables \code{env}, does nothing if \code{if_expr} evaluates
to false and otherwise evaluates \code{run_expr} transparently. If shell is not \code{callr} then the \code{run_expr} gets
wrapped in a \code{system()} command to execute the shell.
}
