% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_main.R
\name{job_maybe_run}
\alias{job_maybe_run}
\title{job_maybe_run}
\usage{
job_maybe_run(job)
}
\arguments{
\item{job}{a named list of job parameters, i.e. one row of the flows table}
}
\value{
invisibly
}
\description{
check to see if it's time to run a job and call the job runner if it is.
the job will be run if following conditions are all true:
\itemize{
\item \code{runs-on} must match the computer name as listed in \code{Sys.info()["nodename"]}
\item \code{if} must be true in the current context
\item \code{needs} must have finished running since the last time this ran (with retval = 0 unless \code{if} evaluates to true)
\item the most recent run in \code{scheduled_runs} is after the last actual run time for this job
}
}
