% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_maybe_start}
\alias{job_maybe_start}
\title{job_maybe_start}
\usage{
job_maybe_start(flow_name, job_name)
}
\arguments{
\item{flow_name}{string workflow name}

\item{job_name}{string job name}
}
\value{
invisibly
}
\description{
check to see if it's time to start a job and call the job runner if it is.
the job will be started if following conditions are all true:
\itemize{
\item \code{runs-on} must match the computer name as listed in \code{Sys.info()["nodename"]}
\item \code{if} must be true in the current context
\item \code{needs} must have finished running since the last time this ran (with retval = 0 unless \code{if} evaluates to true)
\item the most recent run in \code{scheduled_runs} is after the last actual run time for this job
}
}
