% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_on_error}
\alias{job_on_error}
\title{job_on_error}
\usage{
job_on_error(
  flow_name,
  job_name,
  error,
  error_handler = tessiflow::error_handler
)
}
\arguments{
\item{flow_name}{string workflow name}

\item{job_name}{string job name}

\item{error}{error condition object}

\item{error_handler}{function, the error handler to call}
}
\value{
invisibly
}
\description{
updates flows table and database, writes to log, and finalizes running session on error
}
