% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_safely_invoke}
\alias{job_safely_invoke}
\title{job_safely_invoke}
\usage{
job_safely_invoke(job, call, ...)
}
\arguments{
\item{job}{list as returned by \link{flows_get_job}}

\item{call}{character name of function to call}

\item{...}{additional arguments passed on to \code{call}}
}
\value{
result of function or NULL
}
\description{
job_safely_invoke
}
