% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_step}
\alias{job_step}
\title{job_step}
\usage{
job_step(flow_name, job_name)
}
\arguments{
\item{flow_name}{string workflow name}

\item{job_name}{string job name}
}
\value{
invisibly
}
\description{
Run the next step in a job
}
