% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_log.R
\name{log_rotate}
\alias{log_rotate}
\title{log_rotate}
\usage{
log_rotate(filename, size = config::get("tessiflow.logsize") \%||\% 1024^2)
}
\arguments{
\item{filename}{string file name of log file to (potentially) rotate}

\item{size}{integer size in bytes, default is 1 megabyte}
}
\value{
nothing, invisibly
}
\description{
Rotate log file \code{filename} into a zip file if its size exceeds \code{size} bytes
}
