% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_log.R
\name{performance_log_open}
\alias{performance_log_open}
\title{performance_log_open}
\usage{
performance_log_open(flows_log_dir = config::get("tessiflow.log"))
}
\arguments{
\item{flows_log_dir}{directory where the SQLite database is stored}
}
\value{
invisible
}
\description{
performance_log_open
}
