% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_log.R
\name{performance_log_update}
\alias{performance_log_update}
\title{performance_log_update}
\usage{
performance_log_update(
  pids = sapply(ps::ps_find_tree("tessiflow-daemon"), ps::ps_pid)
)
}
\arguments{
\item{pids}{integer vector of process ids}
}
\value{
number of rows updated
}
\description{
Add rows to the performance log table in the SQLite database by calling performance_poll
on all processes in the tessiflow-daemon tree
}
