% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_log.R
\name{performance_main}
\alias{performance_main}
\title{performance_main}
\usage{
performance_main()
}
\value{
never
}
\description{
Main peformance logger loop. Calls performance_log_update every 30 seconds
}
