% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_log.R
\name{performance_poll}
\alias{performance_poll}
\title{performance_poll}
\usage{
performance_poll(pid)
}
\arguments{
\item{pid}{integer process id}
}
\value{
integer list of process data
}
\description{
return a list of process data from \verb{python psutil}
}
