% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schtasks_cron.R
\name{schedule_schtasks}
\alias{schedule_schtasks}
\alias{unschedule_schtasks}
\alias{unschedule_crontab}
\title{schedule_schtasks}
\usage{
schedule_schtasks(command, taskname = "tessiflow")

unschedule_schtasks(taskname = "tessiflow")

unschedule_crontab(taskname = "tessiflow")
}
\arguments{
\item{command}{string command line expression}

\item{taskname}{string name of task}
}
\value{
result of system call
}
\description{
creates a job named \code{taskname} that runs every minute using Windows \code{schtasks.exe}
}
\section{Functions}{
\itemize{
\item \code{unschedule_schtasks()}: deletes a job with the name \code{taskname} using Windows \code{schtasks.exe}

\item \code{unschedule_crontab()}: deletes a job named \code{taskname} using *nix \code{crontab}

}}
