% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schtasks_cron.R
\name{script_expr}
\alias{script_expr}
\title{script_expr}
\usage{
script_expr(expr)
}
\arguments{
\item{expr}{to call in an Rscript process}
}
\value{
Rscript command line call
}
\description{
script_expr
}
