% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_upsert.R
\name{sqlite_upsert}
\alias{sqlite_upsert}
\title{sqlite_upsert}
\usage{
sqlite_upsert(table = "jobs", data, con = tessiflow$db)
}
\arguments{
\item{table}{string, right now only \code{jobs} is valid}

\item{data}{data.frame of data to update, rows will be matched on flow_name, job_name, and start_time}

\item{con}{database connection}
}
\value{
invisibly
}
\description{
sqlite_upsert
}
