% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessiflow.R
\name{tessiflow_enable}
\alias{tessiflow_enable}
\alias{tessiflow_disable}
\title{tessiflow_enable}
\usage{
tessiflow_enable()

tessiflow_disable()
}
\value{
invisibly
}
\description{
Schedule a keep-alive job that runs tessiflow_run.
}
\section{Functions}{
\itemize{
\item \code{tessiflow_disable}: Unschedule the tessiflow job that runs tessiflow_run.
}}

\examples{
\dontrun{
tessiflow::tessiflow_enable()
}
}
