% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tessiflow_flows_get}
\alias{tessiflow_flows_get}
\title{tessiflow_flows_get}
\usage{
tessiflow_flows_get(
  hostname = "localhost",
  port = config::get("tessiflow.port")
)
}
\arguments{
\item{flow_name}{}

\item{job_name}{}
}
\description{
tessiflow_flows_get
}
