% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessiflow.R
\name{tessiflow_job_start}
\alias{tessiflow_job_start}
\title{tessiflow_job_start}
\usage{
tessiflow_job_start(flow_name, job_name)

tessiflow_job_start(flow_name, job_name)
}
\arguments{
\item{flow_name}{string flow name}

\item{job_name}{sting job name}
}
\value{
invisibly
}
\description{
Starts a tessiflow job identified by \code{flow_name} and \code{job_name}
}
\section{Functions}{
\itemize{
\item \code{tessiflow_job_start}: Stops a tessiflow job identified by \code{flow_name} and \code{job_name}
}}

\examples{
\dontrun{
tessiflow::tessiflow_job_start("Workflow A","Job 1")
tessiflow::tessiflow_job_stop("Workflow A","Job 1")
}
}
