% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tessiflow_job_stop}
\alias{tessiflow_job_stop}
\title{tessiflow_job_stop}
\usage{
tessiflow_job_stop(
  flow_name,
  job_name,
  hostname = "localhost",
  port = config::get("tessiflow.port")
)
}
\arguments{
\item{flow_name}{}

\item{job_name}{}
}
\description{
tessiflow_job_stop
}
