% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessiflow.R
\name{tessiflow_pid_lock}
\alias{tessiflow_pid_lock}
\alias{tessiflow_pid_unlock}
\title{tessiflow_pid_lock/unlock}
\usage{
tessiflow_pid_lock(flows_log_dir)

tessiflow_pid_unlock(flows_log_dir)
}
\arguments{
\item{flows_log_dir}{character directory name}
}
\value{
nothing invisibly
}
\description{
Files to manage \code{tessiflow.pid} file to ensure only one process is running at a time
}
\section{Functions}{
\itemize{
\item \code{tessiflow_pid_lock()}: Create \code{tessiflow.pid} file

\item \code{tessiflow_pid_unlock()}: Remove \code{tessiflow.pid} file

}}
