% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{tessiflow_report_load}
\alias{tessiflow_report_load}
\alias{tessiflow_report_send}
\title{tessiflow_report}
\usage{
tessiflow_report_load()

tessiflow_report_send()
}
\value{
nothing
}
\description{
tessiflow_report
}
\section{Functions}{
\itemize{
\item \code{tessiflow_report_load}: Pull process data from the SQLite database from the last 24hours.

\item \code{tessiflow_report_send}: Send the flows report
}}

