% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{tessiflow_report_enable}
\alias{tessiflow_report_enable}
\title{tessiflow_report_enable}
\usage{
tessiflow_report_enable()
}
\value{
TRUE if success, FALSE if failure
}
\description{
Install \code{tessiflow.yml} into the \code{tessiflow.d} directory to run the tessiflow report daily at 9am.
}
