% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessiflow.R
\name{tessiflow_start}
\alias{tessiflow_start}
\title{tessiflow_start}
\usage{
tessiflow_start(flows_log_dir = config::get("tessiflow.log"))
}
\arguments{
\item{flows_log_dir}{directory where \code{tessiflow.log} file and \code{tessiflow.pipe} input will be stored}
}
\value{
the process object
}
\description{
Start a background tessiflow process
}
