% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_parse.R
\name{test_parse}
\alias{test_parse}
\title{test_parse}
\usage{
test_parse(expr_string)
}
\arguments{
\item{expr_string}{expression as a deparsed string}
}
\value{
TRUE if string is parseable
}
\description{
tests if it's possible to parse the expression string, throws an error if it's not
}
