% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_parse.R
\name{test_parse_run}
\alias{test_parse_run}
\title{test_parse_run}
\usage{
test_parse_run(step)
}
\arguments{
\item{expr_string}{expression as a deparsed string}
}
\value{
TRUE if string is parseable
}
\description{
tests if it's possible to parse the step run statement, throws an error if it's not
}
