% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\alias{tessiflow_job_start}
\alias{tessiflow_job_stop}
\alias{tessiflow_flows_get}
\title{Tessiflow API}
\usage{
tessiflow_job_start(
  flow_name,
  job_name,
  hostname = "localhost",
  port = config::get("tessiflow.port")
)

tessiflow_job_stop(
  flow_name,
  job_name,
  hostname = "localhost",
  port = config::get("tessiflow.port")
)

tessiflow_flows_get(
  hostname = "localhost",
  port = config::get("tessiflow.port")
)
}
\arguments{
\item{flow_name}{character workflow name}

\item{job_name}{character job name}

\item{hostname}{hostname of tessiflow server, defaults to \code{localhost}}

\item{port}{port of tessiflow server, defaults to \code{tessiflow.port} config value}
}
\value{
\itemize{
\item \code{tessiflow_job_start()}: list of started job data, as returned by \link{flows_log_get_last_run}
}

\itemize{
\item \code{tessiflow_job_stop()}: list of stopped job data, as returned by \link{flows_log_get_last_run}
}

\itemize{
\item \code{tessiflow_flows_get()}: data.table of all latest job data, as returned by \link{flows_log_get_last_run}
}
}
\description{
The Tessiflow API makes it possible to communicate with a running tessiflow instance
}
\section{Functions}{
\itemize{
\item \code{tessiflow_job_start()}: start job on the tessiflow server

\item \code{tessiflow_job_stop()}: stop job on the tessiflow server

\item \code{tessiflow_flows_get()}: get information on the latest run of each job on the tessiflow server

}}
