% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_flow_job_name.R
\name{check_flow_job_name}
\alias{check_flow_job_name}
\alias{assert_flow_job_name}
\alias{test_flow_job_name}
\title{check_flow_job_name}
\usage{
check_flow_job_name(flow_name, job_name)

assert_flow_job_name(
  flow_name,
  job_name,
  .var.name = checkmate::vname(flow_name),
  add = NULL
)

test_flow_job_name(flow_name, job_name)
}
\arguments{
\item{flow_name}{string flow name}

\item{job_name}{string job name}
}
\description{
check, assertion, and test for flow and job name: both are characters of length 1 and both exist in flows
}
\section{Functions}{
\itemize{
\item \code{assert_flow_job_name()}: errors if check is false

\item \code{test_flow_job_name()}: returns \code{FALSE} if check is false

}}
