% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handlers.R
\name{error_handler}
\alias{error_handler}
\alias{error_print}
\alias{error_email}
\alias{error_calling_handler}
\title{Error handlers}
\usage{
error_handler(error)

error_print(error)

error_email(error)

error_calling_handler(error)
}
\arguments{
\item{error}{error caught}
}
\description{
functions for handling errors
}
\details{
error_handler
}
\section{Functions}{
\itemize{
\item \code{error_handler()}: selects error handler to use

\item \code{error_print()}: prints an error message to the console

\item \code{error_email()}: sends an error message to an email address

\item \code{error_calling_handler()}: global error condition handler, intended to be used with \code{rlang::try_fetch} because it both adds trace information (as a calling handler)
and does error logging (as an error handler)

}}
