% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_start}
\alias{job_start}
\title{job_start}
\usage{
job_start(flow_name, job_name)
}
\arguments{
\item{flow_name}{string workflow name}

\item{job_name}{string job name}
}
\value{
invisibly
}
\description{
Run the job in a callr process
}
