% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_start}
\alias{job_start}
\alias{job_step}
\alias{job_on_error}
\alias{job_poll}
\alias{job_read}
\alias{job_finalize}
\alias{job_reset}
\alias{job_stop}
\title{job_start}
\usage{
job_start(flow_name, job_name)

job_step(flow_name, job_name)

job_on_error(flow_name, job_name, error)

job_poll(flow_name, job_name)

job_read(flow_name, job_name, all = FALSE)

job_finalize(flow_name, job_name)

job_reset(flow_name, job_name)

job_stop(flow_name, job_name)
}
\arguments{
\item{flow_name}{string workflow name}

\item{job_name}{string job name}

\item{error}{error condition object}

\item{all}{boolean whether to wait for the process to finish and return all output. (Most useful when process has already finished.)}
}
\value{
invisibly

character vector of output from process. Names are one or more of \code{output}, \code{error} and \code{process} and match the names from \code{processx::poll_io}
}
\description{
Run the job in a \code{callr} process
}
\section{Functions}{
\itemize{
\item \code{job_step()}: Run the next step in a job

\item \code{job_on_error()}: Updates flows table and database, writes to log, and finalizes running session on error

\item \code{job_poll()}: poll the process \code{stdout} and \code{stderr} streams and call job_step when ready.

\item \code{job_read()}: Read \code{stdout} and \code{stderr} from the process and write to log. When ready, call job_step

\item \code{job_finalize()}: Closes R session, writes to log, console and database, and updates \code{retval}.

\item \code{job_reset()}: Resets job for next run, updates flows table but NOT database, writes to log

\item \code{job_stop()}: Stops job, updates flows table and database, writes to log

}}
