% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{send_email}
\usage{
send_email(
  subject,
  body,
  emails = config::get("tessiflow.email"),
  smtp = config::get("tessiflow.smtp")
)
}
\arguments{
\item{subject}{string subject line}

\item{body}{string body of email}

\item{emails}{list of email addresses, default is from \code{tessiflow.email} configuration variable; \code{from} address will be the first email}

\item{smtp}{list of \code{SMTP} configuration data for \code{\link[mailR:send.mail]{mailR::send.mail}}}
}
\description{
Wrapper for \code{\link[mailR:send.mail]{mailR::send.mail}} that does input checking and pulls in configuration data from the \code{config.yml} file
}
